[feat] Add seeders and factories. Correct navbar.
Separeted the seeders into their separate files. Created a factory for the
Tasks, allowing for dummy task creation and better view at the website.
The header/navbar received a position: absolute, and the div receiving the
page content aquired a padding-top, for the content to not be hidden
behind the header.